Instruction Page
Add a class InstructionPage to display the instructions on how to play the game.
Add a new scene to switch from the game.
Add a back button to switch back to the game.

Add a button in Main to switch to instruction page.
Set all the button size to make it consistent(load,newGame,instruction).
Add the instruction button to root.
Make the instruction button invisible when the game start.
Create a method switchToInstructionPage. When the instruction button is pressed, trigger the method.