Main Menu Page
Split the game scene with the main menu scene.
Create a MainMenuPage Class for the start scene. Migrate the instruction button to this page. Create a start button to switch to game scene.
Create a css file main-menu for the styles of main menu button and main menu images.
Add hitButtonSound in SoundEffects to produce sound when clicking buttons in main menu.

In Main, add a method switchToMainMenuPage.
Modify the start stage to switchToMainMenuPage.
Create a method to initialize the game which is originally in start stage.
The instruction button is moved to the main menu page. Delete the instruction in the root.
Change the start(primaryStage) in loadGame to initializeGame().