Small Changes
Delete bg.jpg in resources. (Not used)
Add a main-menu folder in resources. Move the main-menu related images and css into the folder.
Change the hitButton.m4a to another sound file.
Add background images for in game and gold root. Add changes in style.css.
In Main, avoid the stage to be resized.
In GameEngine, remove the (int) in setFps.