Pause, Restart, Score
Add boolean to determine paused or not. Implement Case P to pause game. Implement method togglePause to pause and resume the game.

Add variables to save the level, heart and score at the start of the level.
In nextLevel method, save the heart and score variable.
Change Case R to restart level.
Change the restart Game to return to initializeNewGame method.
Set pause condition in onPhysicsUpdate.
Implement restartLevel method to restart the certain level.
Change in instruction page that r is for restart level.

In Score, remove the label from the root after it is displayed.