Several Changes
Change title from Game to Brick Ball Game.

In Block, modify the condition of checkHitToBlock for smooth hitting and breaking blocks.

Change instruction, remove load and add m key for volume control. Import images for the label.

Change scene to more specific word, gameScene.

Implement Pause Menu while p is pressed.
Create buttons for resume, restart level and main menu. Add styles with css.

Add a boolean in initializeNewGame method to determine if the game is switch from main menu. If yes, then start from level one. Modify the respective methods.

While win, clear the root to display the word. Add a mainMenu button to switch back to mainMenu.

While switchToInstructionPage, clear the root if not null.

In the handle, remove case L for load. Pass gameScene to togglePause method to implement Pause Menu. Set Pause Menu to visible or invisible while P is pressed.

Change the random yBall initialize position to sceneHeight-200 to avoid the ball to be initializing out of screen.

Add getter method for level, heart and score in Main. To be used in PauseMenu to restartLevel.