LoadSaveManager
Implement a classLoadSaveManager to handle load and save in it.
Change the private variables to protected to be accessed in the class.
Create a LoadSaveManager in the start stage.

Add conditions of heart, score and level when enter new game. If the root is not null, clear the root while initializing the game.
Set the pause condition to false.

Set the back button to invisible while the game proceed to next level.
Change the loadGame and saveGame to handle with using LoadSaveMAnager.