Level Selection
Add saveScore and saveHeart in LoadSaveManager to save the variables while restart from a load game.

Add a levelSelect button in Main. If the game have a saved progress, player can play the past level but with default score and heart. Avoid player to earn extra marks from this.

Add a switchToLevelSelectionPage method to switch scene.

Add a levelSelectionClass.
Create a VBox and HBox to adjust the alignment of the level two in a row.
Create a container VBox to hold the header and buttons.
Create a back button to back to main menu.
Create method playLevel to start playing specific level based on selection.