High Score Feature
Add button in MainMenu to switch to high score page.
While pressing the button, play button click sound.

Move the savePath and savePathDir from main to the loadSaveManager.
Create a highScoreManager in Main to call method checkAndAddHighScore during gameOver and Win.
Add a method switchToHighScorePAge to switch.

Create another constructor in LoadSaveManager to use the class without passing the main.

Add images and css to be used in folder high-score.
Create HighScore class to access the Player name and score.
Create a HighScoreManager class to save and load high scores.
Method updateHighScore save the new score in the highScores and sort based on the score in descending order.
Then, it removes the scores which is not the top 3 and save the new list.
Method saveHighScores output the object into the file specified.
Method loadHighScores check if the file exist and create one if not.
If the file exists then save the object list in highScores.
Method checkAndAddHighScore display the dialog box for player to enter name if achieve high score. If the user did not give input then will not save anything.
Method getFormattedString return thr name and score to be displayed.

The HighScorePage class create a scene to display the marks.
Create button to switch back to main menu.
Get the scores in highscores.mdds.
If no records then did not show anything.
The page is designed with text with label image 1,2,3.
Using Map to join the string and image.