Refactor
Pass gamestate to endgamedisplay.

Refactor block. Move the getpaddingh, getwidth, getheight, getpaddingtop into a subclass BlockSizeGetter. Only for getting the block size.

Created an interface BlockFiller that defines a method applyFill, and concrete implementations for different block types, such as ChocoBlockFiller, HeartBlockFiller, BoomBlockFiller, StarBlockFiller, and NormalBlockFiller. The BlockFillerFactory class is responsible for creating instances of the appropriate BlockFiller based on the block type.

Clear the bonus and boom after gameover or gamewin.