Refactor
Refactor initializeNewGame, onUpdate, onPhysicsUpdate.
Each logical block of code is extracted into a separate method with a descriptive name.
The readability of the code is improved by breaking it into smaller, focused methods.

In onPhysicsUpdate, the common logic for handling power-ups is extracted into a generic method (updatePowerUps).
The specific handling for chocos and booms is parameterized.

Delete unused fps.

Change the root and engine in main to be private. Use getter to access outside the class.